Squishing some warnings.

+++ Src/Game/Entities/BlockEntity.cpp
@@ -78,14 +78,14 @@ void BlockEntity::Init()
 		break;
 	}
 
-	float s_CenterX = (m_GridColumns - 1.0) / 2.0;
-	float s_CenterY = (m_GridRows - 1.0) / 2.0;
+	float s_CenterX = (m_GridColumns - 1.f) / 2.f;
+	float s_CenterY = (m_GridRows - 1.f) / 2.f;
 
-	m_Model->Translate(glm::vec3((m_X - s_CenterX) * 0.30, 0.0, (m_Y - s_CenterY) * 0.30));
+	m_Model->Translate(glm::vec3((m_X - s_CenterX) * 0.3f, 0.0, (m_Y - s_CenterY) * 0.3f));
 
-	m_TargetX = (m_X - s_CenterX) * 0.30;
-	m_TargetY = 0.0;
-	m_TargetZ = (m_Y - s_CenterY) * 0.30;
+	m_TargetX = (m_X - s_CenterX) * 0.3f;
+	m_TargetY = 0.f;
+	m_TargetZ = (m_Y - s_CenterY) * 0.3f;
 
 	Managers::ModelManager::GetInstance()->RegisterModel(m_Model);
 }
@@ -95,7 +95,7 @@ void BlockEntity::Update(double p_Delta)
 	if (m_ScaleStep > 0)
 	{
 		m_Model->Scale(glm::vec3(1.0 - (m_ScaleStep * p_Delta), 1.0 - (m_ScaleStep * p_Delta), 1.0 - (m_ScaleStep * p_Delta)));
-		m_ScaleTimer += p_Delta;
+		m_ScaleTimer += (float) p_Delta;
 
 		if (m_ScaleTimer >= 1.f / m_ScaleStep)
 		{
@@ -108,7 +108,7 @@ void BlockEntity::Update(double p_Delta)
 	{
 		if (m_MovementDelay > 0.f)
 		{
-			m_MovementDelay -= p_Delta;
+			m_MovementDelay -= (float) p_Delta;
 
 			if (m_MovementDelay < 0.f)
 				m_MovementDelay = 0.f;
@@ -124,7 +124,7 @@ void BlockEntity::Update(double p_Delta)
 			if (s_Position.x < m_TargetX)
 			{
 				s_Animating = true;
-				s_Animation.x += 1.5 * p_Delta;
+				s_Animation.x += 1.5f * (float) p_Delta;
 
 				if (s_Position.x + s_Animation.x > m_TargetX)
 					s_Animation.x = m_TargetX - s_Position.x;
@@ -133,7 +133,7 @@ void BlockEntity::Update(double p_Delta)
 			if (s_Position.x > m_TargetX)
 			{
 				s_Animating = true;
-				s_Animation.x -= 1.5 * p_Delta;
+				s_Animation.x -= 1.5f * (float) p_Delta;
 
 				if (s_Position.x - s_Animation.x < m_TargetX)
 					s_Animation.x = m_TargetX - s_Position.x;
@@ -142,7 +142,7 @@ void BlockEntity::Update(double p_Delta)
 			if (s_Position.y < m_TargetY)
 			{
 				s_Animating = true;
-				s_Animation.y += 1.5 * p_Delta;
+				s_Animation.y += 1.5f * (float) p_Delta;
 
 				if (s_Position.y + s_Animation.y > m_TargetY)
 					s_Animation.y = m_TargetY - s_Position.y;
@@ -151,7 +151,7 @@ void BlockEntity::Update(double p_Delta)
 			if (s_Position.y > m_TargetY)
 			{
 				s_Animating = true;
-				s_Animation.y -= 1.5 * p_Delta;
+				s_Animation.y -= 1.5f * (float) p_Delta;
 
 				if (s_Position.y - s_Animation.y < m_TargetY)
 					s_Animation.y = m_TargetY - s_Position.y;
@@ -160,7 +160,7 @@ void BlockEntity::Update(double p_Delta)
 			if (s_Position.z < m_TargetZ)
 			{
 				s_Animating = true;
-				s_Animation.z += 1.5 * p_Delta;
+				s_Animation.z += 1.5f * (float) p_Delta;
 
 				if (s_Position.z + s_Animation.z > m_TargetZ)
 					s_Animation.z = m_TargetZ - s_Position.z;
@@ -169,7 +169,7 @@ void BlockEntity::Update(double p_Delta)
 			if (s_Position.z > m_TargetZ)
 			{
 				s_Animating = true;
-				s_Animation.z -= 1.5 * p_Delta;
+				s_Animation.z -= 1.5f * (float) p_Delta;
 
 				if (s_Position.z - s_Animation.z < m_TargetZ)
 					s_Animation.z = m_TargetZ - s_Position.z;
@@ -193,7 +193,7 @@ void BlockEntity::Update(double p_Delta)
 	{
 		if (m_MovementDelay > 0.f)
 		{
-			m_MovementDelay -= p_Delta;
+			m_MovementDelay -= (float) p_Delta;
 
 			if (m_MovementDelay < 0.f)
 				m_MovementDelay = 0.f;
@@ -202,11 +202,11 @@ void BlockEntity::Update(double p_Delta)
 		{
 			m_PendingPosition = false;
 
-			float s_CenterX = (m_GridColumns - 1.0) / 2.0;
-			float s_CenterY = (m_GridRows - 1.0) / 2.0;
+			float s_CenterX = (m_GridColumns - 1.f) / 2.f;
+			float s_CenterY = (m_GridRows - 1.f) / 2.f;
 
-			float s_TargetX = (m_PendingX - s_CenterX) * 0.30;
-			float s_TargetZ = (m_PendingY - s_CenterY) * 0.30;
+			float s_TargetX = (m_PendingX - s_CenterX) * 0.3f;
+			float s_TargetZ = (m_PendingY - s_CenterY) * 0.3f;
 
 			m_Model->Position(glm::vec3(s_TargetX, m_TargetY, s_TargetZ));
 			m_Model->SetScale(glm::vec3(1.f, 1.f, 1.f));
@@ -277,11 +277,11 @@ void BlockEntity::Position(int p_X, int p_Y, bool p_Simulated, float p_Delay)
 	m_Y = p_Y;
 	m_X = p_X;
 
-	float s_CenterX = (m_GridColumns - 1.0) / 2.0;
-	float s_CenterY = (m_GridRows - 1.0) / 2.0;
+	float s_CenterX = (m_GridColumns - 1.f) / 2.f;
+	float s_CenterY = (m_GridRows - 1.f) / 2.f;
 
-	m_TargetX = (p_X - s_CenterX) * 0.30;
-	m_TargetZ = (p_Y - s_CenterY) * 0.30;
+	m_TargetX = (p_X - s_CenterX) * 0.3f;
+	m_TargetZ = (p_Y - s_CenterY) * 0.3f;
 
 	m_MovementDelay = p_Delay;
